+++ src/app/Console/Commands/Upgrade.php
@@ -10,7 +10,7 @@ class Upgrade extends Command
 
     use PrettyCommandOutput;
 
-    protected $finish_message = ' Dwfw is already upgraded to the most recent version';
+    protected $finish_message = ' Dwfw is already upgraded to the most recent version'; //Needs to be changed in the upgrade methods
 
     /**
      * Array of methods used for upgrading to the given version
@@ -20,6 +20,7 @@ class Upgrade extends Command
 
     protected $upgrade_methods = [
         '0.10.14' => 'upgrade_to_0_10_14',
+        '0.10.15' => 'upgrade_to_0_10_15',
     ];
 
     protected $progressBar;
@@ -47,16 +48,6 @@ class Upgrade extends Command
      */
     public function handle()
     {
-        $this->progressBar = $this->output->createProgressBar(5);
-        $this->progressBar->minSecondsBetweenRedraws(0);
-        $this->progressBar->maxSecondsBetweenRedraws(120);
-        $this->progressBar->setRedrawFrequency(1);
-
-        $this->progressBar->start();
-
-        $this->info(' DWFW upgrade started. Please wait...');
-        $this->progressBar->advance();
-
         foreach ($this->upgrade_methods as $version => $upgrade) {
             if ($version > config('dwfw.version') ?? '0.10.13') {
                 if (is_callable([$this, $upgrade])) {
@@ -67,16 +58,30 @@ class Upgrade extends Command
                 }
             }
         }
-
-        $this->progressBar->finish();
         $this->info($this->finish_message);
+    }
 
+    private function start_progress_bar($version, int $max = 5, int $min_sec_between_redraws = 0, int $max_sec_between_redraws = 120, int $redraw_frequency = 1)
+    {
+        $this->progressBar = $this->output->createProgressBar($max);
+        $this->progressBar->minSecondsBetweenRedraws($min_sec_between_redraws);
+        $this->progressBar->maxSecondsBetweenRedraws($max_sec_between_redraws);
+        $this->progressBar->setRedrawFrequency($redraw_frequency);
+
+        $this->progressBar->start();
+
+        $this->info(' Upgrading to '.$version.'. Please wait...');
+        $this->progressBar->advance();
+    }
+
+    private function finish_progress_bar()
+    {
+        $this->progressBar->finish();
     }
 
     private function upgrade_to_0_10_14()
     {
-
-        $this->info(' Upgrading to 0.10.14');
+        $this->start_progress_bar('0.10.14', 5);
 
         $this->line(' Publishing CheckIpMiddleware');
         $this->executeArtisanProcess('vendor:publish', [
@@ -98,8 +103,28 @@ class Upgrade extends Command
             '--tag' => 'config.dwfw',
             '--force' => '--force',
         ]);
-        $this->finish_message=' DWFW succesfully upgraded. New version: 0.10.14';
+        $this->finish_message = ' DWFW succesfully upgraded. New version: 0.10.14';
         $this->info(' DWFW upgrade to 0.10.14 finished.');
+        $this->finish_progress_bar();
+    }
+
+    private function upgrade_to_0_10_15()
+    {
+        $this->start_progress_bar('0.10.15', 3);
+        $this->line(' Publishing new version number');
+        $this->executeArtisanProcess('vendor:publish', [
+            '--provider' => 'Different\Dwfw\DwfwServiceProvider',
+            '--tag' => 'config.dwfw',
+            '--force' => '--force',
+        ]);
+        $this->line(' Updating Backpack translations');
+        $this->executeArtisanProcess('vendor:publish', [
+            '--provider' => 'Different\Dwfw\DwfwServiceProvider',
+            '--tag' => 'backpack.langs',
+            '--force' => '--force',
+        ]);
+        $this->finish_message = ' DWFW succesfully upgraded. New version: 0.10.15';
+        $this->finish_progress_bar();
     }
 
 }